+++ app-misc/livebuild-scripts/files/livebuild64_xfce.sh
@@ -133,13 +133,13 @@ if  [ -e $BUILDDATA/.work.session ] ; [ -a $LOOP_IMAGE ] ; then
     echo -e "Tarballs exists. \n"
     echo "Session exist"
     exit 1
-else 
+elif [ ! -e $BUILDDATA/stage3-amd64.tar.bz2 ] ; [ ! -a $BUILDDATA/portage-latest.tar.xz ] ; then
     echo -e "Downloading Stage3"
     sleep 1
     wget -c -O $BUILDDATA/stage3-amd64.tar.bz2 $STAGE3
     echo -e "Downloading Portage"
     sleep 1
-    wget -c -O $BUILDDATA/portage-latest.tar.bz2 $PORTS
+    wget -c -O $BUILDDATA/portage-latest.tar.xz $PORTS
     echo -e "Create block device \n"
     sleep 1
     dd if=/dev/zero of=$LOOP_IMAGE bs=1024K count=20000  || die() { echo "$@" 1>&2 ; exit 1; }
@@ -153,7 +153,7 @@ else
     cp -L /etc/resolv.conf $BUILDROOT/etc/
     echo -e "Unpacking Portage \n"
     sleep 1
-    tar xpf $BUILDDATA/portage*.tar.bz2 -C $BUILDROOT/usr 2> $BUILDROOT/buildport.log
+    tar xpf $BUILDDATA/portage*.tar.xz -C $BUILDROOT/usr 2> $BUILDROOT/buildport.log
     touch $BUILDDATA/.work.session
     echo -e "Stage3 & Portage Unpacked \n" 
     sleep 1
